Introduced a `resolver_repo.go`
which holds all the methods related to resolving a package in a repo. This is a minor followup cleanup PR to #1047.
