+++ pkg/kudoctl/util/repo/repo_operator.go
@@ -1,18 +1,13 @@
 package repo
 
 import (
-	"bytes"
 	"fmt"
 	"io/ioutil"
 	"net/url"
 	"strings"
 
-	"github.com/kudobuilder/kudo/pkg/kudoctl/clog"
 	"github.com/kudobuilder/kudo/pkg/kudoctl/http"
 	"github.com/kudobuilder/kudo/pkg/kudoctl/kudohome"
-	"github.com/kudobuilder/kudo/pkg/kudoctl/packages"
-	"github.com/kudobuilder/kudo/pkg/kudoctl/packages/reader"
-
 	"github.com/pkg/errors"
 	"github.com/spf13/afero"
 )
@@ -75,72 +70,4 @@ func (c *Client) DownloadIndexFile() (*IndexFile, error) {
 
 	indexFile, err := ParseIndexFile(indexBytes)
 	return indexFile, err
-}
-
-// getPackageReaderByAPackageURL downloads the tgz file from the remote repository and returns a reader
-// The PackageVersion is a package configuration from the index file which has a list of urls where
-// the package can be pulled from.  This will cycle through the list of urls and will return the reader
-// from the first successful url.  If all urls fail, the last error will be returned.
-func (c *Client) getPackageReaderByAPackageURL(pkg *PackageVersion) (*bytes.Buffer, error) {
-	var pkgErr error
-	for _, u := range pkg.URLs {
-		r, err := c.getPackageBytesByURL(u)
-		if err == nil {
-			return r, nil
-		}
-		pkgErr = fmt.Errorf("unable to read package %w", err)
-		clog.V(2).Printf("failure against url: %v  %v", u, pkgErr)
-	}
-	clog.Printf("Giving up with err %v", pkgErr)
-	return nil, pkgErr
-}
-
-func (c *Client) getPackageBytesByURL(packageURL string) (*bytes.Buffer, error) {
-	clog.V(4).Printf("attempt to retrieve package from url: %v", packageURL)
-	resp, err := c.Client.Get(packageURL)
-	if err != nil {
-		return nil, errors.Wrap(err, "getting package url")
-	}
-
-	return resp, nil
-}
-
-// GetPackageBytes provides an io.Reader for a provided package name and optional version
-func (c *Client) GetPackageBytes(name string, version string) (*bytes.Buffer, error) {
-	clog.V(4).Printf("getting package reader for %v, %v", name, version)
-	clog.V(5).Printf("repository using: %v", c.Config)
-	// Construct the package name and download the index file from the remote repo
-	indexFile, err := c.DownloadIndexFile()
-	if err != nil {
-		return nil, errors.WithMessage(err, "could not download repository index file")
-	}
-
-	pkgVersion, err := indexFile.GetByNameAndVersion(name, version)
-	if err != nil {
-		return nil, errors.Wrapf(err, "getting %s in index file", name)
-	}
-
-	return c.getPackageReaderByAPackageURL(pkgVersion)
-}
-
-// Resolve provides an Package for a provided package name and optional version
-func (c *Client) Resolve(name string, version string) (*packages.Package, error) {
-	buf, err := c.GetPackageBytes(name, version)
-	if err != nil {
-		return nil, err
-	}
-	files, err := reader.ParseTgz(buf)
-	if err != nil {
-		return nil, err
-	}
-
-	resources, err := files.Resources()
-	if err != nil {
-		return nil, err
-	}
-
-	return &packages.Package{
-		Resources: resources,
-		Files:     files,
-	}, nil
 }
